requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
authlib==1.3.0
requests==2.31.0
google-generativeai==0.3.2
